requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
httpx>=0.27.0
anthropic>=0.40.0
composio-core>=0.6.0
composio-anthropic>=0.6.0
